rename physical_approximations to external_generated_inputs
