+++ api/osrd_infra/models/__init__.py
@@ -4,4 +4,4 @@ from osrd_infra.models.path import *  # noqa
 from osrd_infra.models.rolling_stock import *  # noqa
 from osrd_infra.models.timetable import *  # noqa
 from osrd_infra.models.train_schedule import *  # noqa
-from osrd_infra.models.physical_approximations import *  # noqa
+from osrd_infra.models.external_generated_inputs import *  # noqa
