+++ api/osrd_infra/schemas/external_generated_inputs.py
